Updated readme
I have attempted to add some information and guidance on the helper methods in the WaitAndRetry project. I've re-stated concepts covered elsewhere in the Polly project, especially around jitter. I've tried to provide some
prescriptive guidance for people who might not know when to use which method or what defaults might look like. That said, my guidance could be very off! So, I'd be grateful for feedback there.

There is no guideline on contributing. If that's necessary, maybe just a pointer back to the Polly readme is sufficient?

I've chosen to use `.WaitAndRetryAsync()` for all policy usage to focus on the wait-and-retry helpers. However, I'm worried that may add confusion and cause the reader to conclude that this will not work with `.WaitAndRetry()`. I
can either change the first example to use `WaitAndRetry()` or add a section explictly addressing this. Or, leave it as-is.